refactor: 커스텀 훅 엔트리 파일 추가 및 경로 설정 수정
